refactor: 네비게이션 패키지 생성
