+++ core/common/src/main/java/com/kitching/core/common/widget/BottomNavItem.kt
@@ -6,7 +6,7 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.graphics.vector.ImageVector
 import androidx.compose.ui.res.vectorResource
 import com.kitching.core.common.R
-import com.kitching.core.common.ScreenRouteDef
+import com.kitching.core.common.navigation.ScreenRouteDef
 
 data class BottomNavItem(
     val tabName: String = "",
