+++ core/navigation/src/main/java/com/kitching/core/navigation/AppNavHost.kt
@@ -8,7 +8,7 @@ import androidx.compose.ui.Modifier
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
-import com.kitching.core.common.ScreenRouteDef
+import com.kitching.core.common.navigation.ScreenRouteDef
 import com.kitching.core.common.commonstate.rememberCommonState
 import com.kitching.login.ui.screen.splash.SplashScreen
 import com.kitching.login.navigation.loginNavGraph
