+++ feature/main/src/main/java/com/kitching/main/navigation/ChattingNavGraph.kt
@@ -3,7 +3,7 @@ package com.kitching.main.navigation
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.compose.composable
 import androidx.navigation.navigation
-import com.kitching.core.common.ScreenRouteDef
+import com.kitching.core.common.navigation.ScreenRouteDef
 import com.kitching.core.common.commonstate.CommonState
 import com.kitching.main.ChatTabScreen
