+++ feature/main/src/main/java/com/kitching/main/navigation/CustomNavHost.kt
@@ -7,7 +7,7 @@ import androidx.compose.ui.Modifier
 import androidx.navigation.NavHostController
 import androidx.navigation.compose.NavHost
 import com.kitching.core.common.commonstate.CommonState
-import com.kitching.core.common.ScreenRouteDef
+import com.kitching.core.common.navigation.ScreenRouteDef
 
 @Composable
 fun CustomNavHost(
