+++ feature/main/src/main/java/com/kitching/main/navigation/MainNavGraph.kt
@@ -5,7 +5,7 @@ import androidx.navigation.NavHostController
 import androidx.navigation.compose.composable
 import androidx.navigation.navigation
 import com.kitching.core.common.commonstate.CommonState
-import com.kitching.core.common.ScreenRouteDef
+import com.kitching.core.common.navigation.ScreenRouteDef
 import com.kitching.main.EntryPointScreen
 
 fun NavGraphBuilder.mainNavGraph(
