+++ feature/main/src/main/java/com/kitching/main/navigation/OtherNavGraph.kt
@@ -4,7 +4,7 @@ import androidx.navigation.NavGraphBuilder
 import androidx.navigation.NavHostController
 import androidx.navigation.compose.composable
 import androidx.navigation.navigation
-import com.kitching.core.common.ScreenRouteDef
+import com.kitching.core.common.navigation.ScreenRouteDef
 import com.kitching.core.common.commonstate.CommonState
 
 fun NavGraphBuilder.otherNavGraph(commonState: CommonState, navController: NavHostController) {
