+++ feature/main/src/main/java/com/kitching/main/navigation/ScheduleNavGraph.kt
@@ -6,7 +6,7 @@ import androidx.navigation.NavType
 import androidx.navigation.compose.composable
 import androidx.navigation.navArgument
 import androidx.navigation.navigation
-import com.kitching.core.common.ScreenRouteDef
+import com.kitching.core.common.navigation.ScreenRouteDef
 import com.kitching.core.common.commonstate.CommonState
 import com.kitching.main.view.schedule.ScheduleDetailScreen
 import com.kitching.main.view.schedule.ScheduleTabScreen
